Restructure into core, client and server.
