+++ core/src/main/java/au/org/ala/util/BasicNormaliser.java
