+++ core/src/main/java/au/org/ala/util/Normaliser.java
