+++ core/src/main/java/au/org/ala/vocab/IsDefinedBy.java
