+++ core/src/test/java/au/org/ala/util/BasicNormaliserTest.java
