+++ server/src/main/resources/groups.json
