+++ server/src/main/resources/subgroups.json
